fix: remove ABORTED status from Routine
This status was used on trigger a pending | scheduled | failed routine to revert.
The revert flow might be very complex, in order to try make it easier this commit remove this status.

From now when asking to revert a routine the status will be changed to "reverting" then we can follow the
basic revert flow