fix: ensure that the Routine retry flow is properly dispatched
We are not able to start the retry flow of a pipeline because the Routine._diff was not being properly calculated. To fix that, I just ensure that the ModelDiffMixin.save will be triggered by changing the inheritance ordering.

This commit also improves some tests in order to ensure that the retrying flow (failure and success) is being properly tested.