fix: subscriber initialization should not assume all tasks follow the default naming convention
I've noticed, at Nilo, that all our deploys think the catalog subscriber has not been added yet:

```
make prd-register_subscriptions
./almanac_api/manage.py initialize_subscribers
Successfully initialized 3 subscribers to domain https://almanac-api-xxxxxxxxxxxx.us-east1.run.app
- [+] CatalogAutoGeneratorSubscriber
- [~] CareProviderSubscriberTask
- [~] RefereeInfoSubscriberTask
```

That is because this subscription uses a target HTTP endpoint that does not follow the default conventions. django-cloud-tasks allow customizing this, so it is not like the library is being used in an unexpected way.

The problem, however, is not adding the subscriber, but rather removing it: since it is never found, once it is no longer offered, the subscription won't be erased.

This commits change the intiailize_subscribers strategy so that it matches subscriptions by domain, and not by name